Split theme into ANSI palette and pi-tui adapters
Move markdownTheme/selectListTheme/editorTheme into
renderers/pi-tui/theme-adapters.ts. theme.ts now holds only the renderer-agnostic
ANSI palette (the Theme class + tokens); the pi-tui-shaped theme objects are built
inside the renderer. components.ts, hooks.ts and frontend.ts import the adapters
from the new location.

Continues the substrate/renderer split — a different renderer builds its own
native theme objects from the same palette.
